Moved convolutions and exposed them directly
